+++ iaso/api/org_unit_change_requests/serializers.py
@@ -1,0 +1,62 @@
+from rest_framework import serializers
+
+from iaso.models import OrgUnitChangeRequest
+
+
+class MobileOrgUnitChangeRequestListSerializer(serializers.ModelSerializer):
+    org_unit_id = serializers.IntegerField(source="org_unit.id")
+    org_unit_uuid = serializers.UUIDField(source="org_unit.uuid")
+
+    class Meta:
+        model = OrgUnitChangeRequest
+        fields = [
+            "id",
+            "org_unit_id",
+            "org_unit_uuid",
+            "status",
+            "approved_fields",
+            "rejection_comment",
+            "created_at",
+            "updated_at",
+            "new_parent_id",
+            "new_name",
+            "new_org_unit_type_id",
+            "new_groups",
+            "new_location",
+            "new_accuracy",
+            "new_reference_instances",
+        ]
+
+
+class OrgUnitChangeRequestListSerializer(serializers.ModelSerializer):
+    org_unit_id = serializers.IntegerField(source="org_unit.id")
+    org_unit_uuid = serializers.UUIDField(source="org_unit.uuid")
+    org_unit_name = serializers.CharField(source="org_unit.name")
+    org_unit_type_id = serializers.IntegerField(source="org_unit.org_unit_type.id")
+    org_unit_type_name = serializers.CharField(source="org_unit.org_unit_type.name")
+    groups = serializers.SerializerMethodField(method_name="get_current_org_unit_groups")
+    created_by = serializers.CharField(source="created_by.username", default="")
+    updated_by = serializers.CharField(source="updated_by.username", default="")
+
+    class Meta:
+        model = OrgUnitChangeRequest
+        fields = [
+            "id",
+            "org_unit_id",
+            "org_unit_uuid",
+            "org_unit_name",
+            "org_unit_type_id",
+            "org_unit_type_name",
+            "status",
+            "groups",
+            "requested_fields",
+            "approved_fields",
+            "rejection_comment",
+            "created_by",
+            "created_at",
+            "updated_by",
+            "updated_at",
+        ]
+
+    def get_current_org_unit_groups(self, obj: OrgUnitChangeRequest):
+        return [group.name for group in obj.org_unit.groups.all()]
