+++ iaso/tests/api/org_unit_change_requests/test_org_unit_change_requests.py
@@ -2,7 +2,7 @@ import time_machine
 
 from django.contrib.gis.geos import Point
 
-from iaso.api.org_unit_change_requests import OrgUnitChangeRequestListSerializer
+from iaso.api.org_unit_change_requests.serializers import OrgUnitChangeRequestListSerializer
 from iaso.test import TestCase
 from iaso.test import APITestCase
 from iaso import models as m
